Abstract away details about magnitudes and dampening movement
Subclasses of AbsBoatMovementController no longer have to concern themselves with making a call to dampen movement when necessary.
Optionally, subclasses can assign the value of calculatedMaxMagnitude to override maxMagnitude when desired without interfering with maxMagnitude itself.

Also increased the default assigned maxMagnitude of the player in Awake() to 3.25f to adjust for incerased camera distance.